requirements.txt
python-dotenv==1.2.2
psycopg2==2.9.11
newspaper4k==0.9.5
openai==2.25.0